PS1 Updates
Removed the PROMPT function since we are using a whole file just for the
PS1 variable.  Added some variables to rc for this as well that are
related to /etc/bash_completion.d/git variables needed for __git_ps1()
to function correctly. We are currently using the default git bash
command line completion file rather than version our own.