Home/Work Settings and Plurals
Renamed the bash alias file to alias rather than aliases since there may
only be one alias and not more than one. Added some include for the
previous work and home directories that are ignored by the .gitignore
file. The include function will skip them if they don't exist. The
tricky part is that using the includes from the work directory makes the
paths longer when source those files.

ie.) include bash/work/alias

This is how work include paths need to be formed. They are based off of
the $HOMBASE variable at the top of the bash/rc file.