SSH, Git and Bash
Still working on adding final touches for bash. Mostly based on SSH
things that need to happen. Stubbed in the git color file for colorizing
the PS1 output for bash. Started adding in SSH.  After these are done we
are going start focusing in more on our dotfile repo and reintroducing
this into fresh. We need to get things right with vim and vundel, etc.