Major Clean-up
Created a rc file for .homb and a lib file that has the hyper .homb specific
functions there rather than in the bash specific lib and rc files.

Updated the README.md so it makes more sense and gives the reader a
better understanding of what .homb is and how to quickly install it
and fresh.

Discussion on incorporating fresh's install to homb is being considered
so some of thie may change again with that decision.